PM: add memory barrier after write PMCSR_SLP to PMCSR register
After execute sleep/wakeup for a substantial many times(nearly 200 times),
the hardware can't enter into sleep state, from the LED for sleep indicator
we can see that write PMCSR_SLP to PMCSR register failed, so the kernel is
already at suspend state while the hardware is not at sleep state.
But the reason why it is not execute successfully is unknown.

Add a msync after write PMCSR_SLP to PMCSR register can resolve the problem.
The msync instruction ensures that all instructions preceding the msync
have completed before msync completes and that no subsequent instructions
are initiated until after the msync completes, and although it adds overhead
to the sleep phase, it is insignificant and poses no issues to the system.

Signed-off-by: Xufeng Zhang <xufeng.zhang@windriver.com>