BOOK3E/KDUMP: go to label with LOAD_REG_ADDR via toc for RELOCATABLE kernel
We can't just use a direct branch to label because the distance
from here to there depends on where the kernel ends up being put.
So we have to go anywhere with LOAD_REG_ADDR via toc.

Additionally, for interrupt label we use SPRN_SPRG3/4 to save/restore used GPRGs
since for 64-bit embedded
        - SPRG3 unused
        - SPRG4 unused
its safe enough. And we have to extend the gap between each interrupt because we
need more instructions to implement this.

Signed-off-by: Tiejun Chen <tiejun.chen@windriver.com>