qman: remove CONFIG_FSL_QMAN_NULL_FQ_DEMUX
The current implementation of this feature is complicating other
development work, in part because it was not implemented properly.
Eg. the use of the "struct qman_fq_cb" type implies a default ERN
handler, yet s/w ERN handling by its nature can not be part of NULL
demuxing - enqueues require a FQ object, so that is the object that
the ERN demuxes to, so there is no NULL case. (In particular, ERNs
demux using a "tag" field that is copied from the enqueue command
because there is no "contextB" in this path of the qman interface,
whereas NULL_FQ_DEMUX is all about setting contextB to zero...)

In terms of use-cases, DC_ERN handling is the only aspect of this
feature that is a known requirement, and that case is outright broken
in the current implementation (I have reimplemented DC_ERN handling
via a new interface in a subsequent patch).

Finally, for these kind of handlers (ie. not based on FQ objects) to
be flexible, it should be possible to set them globally (ie. for use
by all portals managed by the driver instance) which the current
implementation doesn't allow. Ie. if users want the same behaviour for
each portal/core, they shouldn't have to repeat the same API call for
all of them. This NULL_FQ_DEMUX feature will be reimplemented properly
once the current rework is done.

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>
Signed-off-by: Kumar Gala <galak@kernel.crashing.org>
[Kevin: Original patch taken from FSL
QorIQ-SDK-V1.2-SOURCE-20120614-yocto.iso image]
Signed-off-by: Kevin Hao <kexin.hao@windriver.com>