qman: handle DCP ERNs per-portal/globally, not per-FQ.
Lookups for DCP ERNs were based on FQ, which is completely broken, as
it turns out. This patch instead allows per-portal and/or global DCP
ERN handlers to be registered, with the per-portal handler taking
precedence if it is set. This change removes the unusable callback
from "struct qman_fq", which may help in certain scenarios where
higher-layer extensions to stashable FQ context were struggling to fit
within a cacheline boundary.

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>
Signed-off-by: Kumar Gala <galak@kernel.crashing.org>
[Kevin: Original patch taken from FSL
QorIQ-SDK-V1.2-SOURCE-20120614-yocto.iso image]
Signed-off-by: Kevin Hao <kexin.hao@windriver.com>