powerpc: don't enable irq in program exception
The commit cd8a5673 (powerpc: Fix might-sleep warning in program
check exception handler) caused a subtle regression in the irq in
program exception handler in the kprobe return from interrupt in that
it enables interrupts when they should not have.  There is a dedicated
stack for program exception when kprobes is enabled, and this requires
that the program exception handler must be nonpreemptive.
The instruction that is being read was actually the faulting
instruction so there should not be any further data faults.
In this case we can replace get_user() with probe_kernel_read()
and keep the IRQs disabled.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>
Reviewed-by: Jason Wessel <jason.wessel@windriver.com>