ARM: kernel: add logical mappings look-up
Commit 32c4ee9271053c0a980c53a5ca8e4526fc95a861 from
git://git.linaro.org/kernel/linux-linaro-tracking.git

In ARM SMP systems the MPIDR register ([23:0] bits) is used to uniquely
identify CPUs.

In order to retrieve the logical CPU index corresponding to a given
MPIDR value and guarantee a consistent translation throughout the kernel, this
patch adds a look-up based on the MPIDR so that irq controller drivers and
other kernel subsystems can use it whenever the logical cpu index corresponding
to a given MPIDR value is needed.

Signed-off-by: Lorenzo Pieralisi <lorenzo.pieralisi@arm.com>
Signed-off-by: Guoqing Jiang <Guoqing.Jiang@windriver.com>