Revert "sdhci: enable cache snooping"
Since DMA cache snooping is already enabled in esdhc_enable_dma(),
so we don't need to enable it again in sdhci_init(). Furthmore,
DMA control register is not an standard register in SDHCI
Specification, kernel may crash when there is not such register
on board.
So Revert the useless commit "sdhci: enable cache snooping".

Signed-off-by: Xufeng Zhang <xufeng.zhang@windriver.com>