Preliminary thread support for e6500
Add support for starting hardware threads on e6500.

The general idea is that each core will release all of its
threads into the secondary thread startup code, which will
eventually wait in the secondary core holding area, for the
appropriate bit in the PACA to be set. The kick_cpu function
pointer will set that bit in the PACA, and thus "release"
the core/thread to boot.

This is my third attempt. It boots with 24 threads. The big
change was the simulator, but I also moved thread enabling so
that only the first thread was doing it.

Caches are probably screwed up, as the "restore" function
doesn't know about the shared nature of the caches.

Also of note: there's no MMU protection. I was hoping that we could
show threads working without robustness, but that may be wrong. :)

Signed-off-by: Andy Fleming <afleming@freescale.com>
[Kevin: Original patch taken from fsl sdk 1.3.1
QorIQ-SDK-V1.3.1-SOURCE-20121220-yocto.iso. Just some minor context
mods in order to port to 3.4 kernel.]
Signed-off-by: Kevin Hao <kexin.hao@windriver.com>