uio: Add FSL RMan UIO driver support
Add support to expose FSL RMan controller to user space.  The RMan is
composed of a global configuration and multiple inbound classification
units.  We map the global configuration as one UIO device and each
classification unit as a unique UIO device.

Signed-off-by: Minghuan Lian <Minghuan.Lian@freescale.com>
Signed-off-by: Kumar Gala <galak@kernel.crashing.org>

[Extract from QorIQ-DPAA-SDK-v1.1-20111026-systembuilder.iso
Fixed context to apply to Wind River kernel tree.]

Integrated-by: Jiang Lu <lu.jiang@windriver.com>