net/dpa: fix get/set pauseparam interfaces for ethtool
Problem:
Failed to use ethtool to set and get the correct status of tx/rx
pause-frames.

Root cause:
The interfaces for ethtool get and set pauseparam is incomplete,
so "Rx_Flow" and "Tx_Flow" bits in MACCFG1 register which reflect
the PAUSE-frames status can't be set correctly and also can't be
got correctly by ethtool.

Fix:
Use the existing interfaces and add some necessary interfaces to
implement this function.

Signed-off-by: Xufeng Zhang <xufeng.zhang@windriver.com>