dpaa_eth: Replaced BUG_ON with WARN_ON
By replacing BUG_ON with WARN_ON the board will be able
to complete the boot process, hence someone that is
working on a different task will not be interrupted
by the failing unit test.

May the unit test fail, the error log can be analyzed by
automatic testing tools as it is a standardized format.

Signed-off-by: Cristian Sovaiala <cristian.sovaiala@freescale.com>
[Kevin: Original patch taken from FSL
QorIQ-SDK-V1.2-SOURCE-20120614-yocto.iso image]
Signed-off-by: Kevin Hao <kexin.hao@windriver.com>