dpaa_eth: use qman_affine_channel() for Rx setup
This removes a hack that was put in before qman_affine_channel() was
available. In order to determine the channel IDs of the cpu-affine
portals, it would parse the device-tree directly and try to guess the
which ones the qman driver would have associated with which cpus.
Changes to the qman driver (and removal of static assumptions from the
device-tree) are completely invalidating those assumptions.

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>
[Kevin: Original patch taken from FSL
QorIQ-SDK-V1.2-SOURCE-20120614-yocto.iso image]
Signed-off-by: Kevin Hao <kexin.hao@windriver.com>