timer/list: pass correct function entry address
Function Descriptor is introduced for PPC64 ABI. It consists of
three words as follows:
1st word: This function real entry point address
2nd word: TOC base address of this function
3rd word: Others

And the function symbol only means this associated function descriptor.

So we should call dereference_function_descriptor to convert that to
the first word as address to lookup symbol name for PPC64 if
pass function symbol.

This should be fine for all arch since dereference_function_descriptor()
is no-op for other arches that don't need this.

Signed-off-by: Tiejun Chen <tiejun.chen@windriver.com>