qbman: do not disable pre-emption in poll APIs
Poll functions assume the caller is cpu-affine and in no risk of
re-entrance, which are the two reasons we usually use the
get/put_cpu_var() semantic - ie. to disable pre-emption. Some use-cases
expect the execution context to remain as non-atomic during
poll-triggered callbacks as it was when the poll API was first called
(eg. NAPI), so we go out of our way in this case to not disable
pre-emption

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>
[Kevin: Original patch taken from FSL
QorIQ-SDK-V1.2-SOURCE-20120614-yocto.iso image]
Signed-off-by: Kevin Hao <kexin.hao@windriver.com>