Handle MPIC error interrupt as a cascaded interrupt.
All SOC device error interrupts are muxed and delivered to the core as a single
MPIC error interrupt. Currently all the device drivers requiring access to device
errors have to register for the MPIC error interrupt as a shared interrupt.

With this patch we add interrupt demuxing capability in the mpic driver, allowing
device drivers to register for their individual error interrupts. This is achieved
by handling error interrupts in a cascaded fashion.

MPIC error interrupt is handled by the "error_int_handler", which subsequently demuxes
it using the EISR and delivers it to the respective drivers.

The error interrupt capability is dependent on the MPIC EIMR register, which was
introduced in MPIC version 4.1 (P4080 rev2). So, error interrupt demuxing capability
is dependent on the MPIC version and can be used for versions >= 4.1.

Signed-off-by: Varun Sethi <Varun.Sethi@freescale.com>
[Kevin: Original patch taken from FSL
QorIQ-SDK-V1.2-SOURCE-20120614-yocto.iso image. Adjust the arguments
of mpic_map and replace MPIC_PRIMARY with MPIC_SECONDARY. Also change
the value of MPIC_FSL_HAS_EIMR to 0x00010000 since there is no
MPIC_NO_READBACK in our kernel.]
Signed-off-by: Kevin Hao <kexin.hao@windriver.com>