powerpc/85xx: PAMU driver API should take a node, not an LIODN
Instead of passing an actual LIODN to the PAMU driver's stashing API,
pass a device tree node and "device index".  Each device tree node for
an LIOD includes an fsl,liodn property that contains an array of LIODNs.
So instead of asking the caller to parse the property, the caller passes
the node itself and an index into the property.

This change eliminates the need for the caller to know the name of the
property.  When we add hypervisor support, a different device tree property
will be used, so we want to hide that from the caller.

Signed-off-by: Timur Tabi <timur@freescale.com>
[Kevin: Original patch taken from FSL
QorIQ-SDK-V1.2-SOURCE-20120614-yocto.iso image]
Signed-off-by: Kevin Hao <kexin.hao@windriver.com>