KVM: Factor out kvm_vcpu_kick to arch-generic code
commit b6d33834bd4e8bdf4a199812e31b3e36da53c794 upstream.

The kvm_vcpu_kick function performs roughly the same funcitonality on
most all architectures, so we shouldn't have separate copies.

PowerPC keeps a pointer to interchanging waitqueues on the vcpu_arch
structure and to accomodate this special need a
__KVM_HAVE_ARCH_VCPU_GET_WQ define and accompanying function
kvm_arch_vcpu_wq have been defined. For all other architectures this
is a generic inline that just returns &vcpu->wq;

Acked-by: Scott Wood <scottwood@freescale.com>
Signed-off-by: Christoffer Dall <c.dall@virtualopensystems.com>
Signed-off-by: Marcelo Tosatti <mtosatti@redhat.com>
Signed-off-by: Avi Kivity <avi@redhat.com>
[tt: removed arch/s390/kvm/kvm-s390.c since we only keep x86 and powerpc kvm]
Signed-off-by: Thomas Tai <thomas.tai@windriver.com>