qbman: don't spawn threads to init portals
Portal-initialisation and recovery-exit needs to execute on the affine
core, but this can be achieved by changing the cpumask of the current
task rather than spawning a thread to do it. The latter, apart from
being overkill, also requires a lot more code.

Also, the return value from [qb]man_recovery_exit() was to convey any
errors associated with thread-creation, which is no longer relevant,
so the API is changed to be void-typed.

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>
[Kevin: Original patch taken from FSL
QorIQ-SDK-V1.2-SOURCE-20120614-yocto.iso image]
Signed-off-by: Kevin Hao <kexin.hao@windriver.com>