qbman: remove UIO support
As with recovery support, UIO too needs to be pulled out in order to
transition the portal initialisation usage to a more dynamic and
modular form. UIO support will be added back in in a way that better
suits the less statically-assigned model we're moving to.

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>
[Kevin:  Original patch taken from FSL
QorIQ-SDK-V1.2-SOURCE-20120614-yocto.iso image. Remove the mods for
the specific board's dts.]
Signed-off-by: Kevin Hao <kexin.hao@windriver.com>