qman: invalidate all DQRR entries on initialisation
If a portal is reinitialised for use on a different CPU to the previous
usage of the portal, there is a possibility that a stale cache entry
could be misinterpreted as a valid entry after reinitialisation, leading
to undefined results. This can't happen if the portal is reused on the
same CPU as the previous usage, because "valid bit" polarity prevents
that happening. The recently added functionality that removes the
assumption of a static portal->cpu assignment is what creates this
theoretical possibility. The simplest solution is to forcibly invalidate
the cachelines for all DQRR entries on portal initialisation.

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>
[Kevin: Original patch taken from FSL
QorIQ-SDK-V1.2-SOURCE-20120614-yocto.iso image]
Signed-off-by: Kevin Hao <kexin.hao@windriver.com>