qbman/dpaa_eth/usdpaa: dynamic pool-channel/CGR allocation
Pool-channels are now described via a range node+property (specifying the
range of channel IDs) rather than a separate node for each pool containing
a channel ID property. The set of valid CGR IDs are also specified via a
range node+property. Allocator APIs are added for both type of object
(in-kernel as well as for USDPAA).

Consequently, the dpaa_eth "ethernet" nodes no longer link to the pool
channel they are to use, the driver is modified to instead dynamically
allocate a single pool-channel for all its interfaces.

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>
[Kevin: Original patch taken from FSL
QorIQ-SDK-V1.2-SOURCE-20120614-yocto.iso image. Remove the mods
for the specific board's dts.]
Signed-off-by: Kevin Hao <kexin.hao@windriver.com>
Signed-off-by: Tiejun Chen <tiejun.chen@windriver.com>
Integrated-by: Yongli He <yongli.he@windriver.com>