netdev/dpa: probe all the specific buffer pool in native bsp
Currently the dpa driver only supports one kernel buffer pool
for all the Ethernet controllers. This works fine when all the
Ethernet use the pool allocated dynamically by the dpa driver
in kernel. But in order to support linux-linux guest os, we have
to use specific pool for each Ethernet. So we introduce a patch
(net/dpa: only use one buffer pool even when static buffer) to
only use one buffer pool even several specific pool are used
for Ethernet. But when an Ethernet is reserved for userspace
driver, we still need to setup the buffer pool for this Ethernet.
So add a check here to distinguish between these two cases.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>