dpaa_eth: Fix incorrect use of DMA mapping direction
Tx buffers were being mapped with DMA_TO_DEVICE flag, which was
incorrect in the case of recycling, when those buffers ended up
back in the buffer pool and were re-used on Rx.

This patch introduces a bidirectional mapping for recycled buffers.
Newly allocated buffers which are used to populate the pool are now
also mapped this way, to avoid any inconsistencies when doing the
unmapping on Rx.

Slight performance decrease in forwarding scenarios (<2%).

Signed-off-by: Ioana Radulescu <ruxandra.radulescu@freescale.com>
[Kevin: Original patch taken from FSL
QorIQ-SDK-V1.2-SOURCE-20120614-yocto.iso image]
Signed-off-by: Kevin Hao <kexin.hao@windriver.com>