dpaa_eth: Add placeholders for fastpath hooks
Added hook points in the datapath dequeue callbacks, which other modules
can use for intercepting FDs/skbs. Actual hooks are not defined (yet),
only a registration function is exported.

This required that the fqid be explicitly passed to internal processing
functions on the datapath (_dpa_rx() and friends).

Signed-off-by: Bogdan Hamciuc <bogdan.hamciuc@freescale.com>
[Kevin: Original patch taken from FSL
QorIQ-SDK-V1.2-SOURCE-20120614-yocto.iso image]
Signed-off-by: Kevin Hao <kexin.hao@windriver.com>