gianfar: Add skb recycling for nonlinear skbs
Enable skb recycling for nonlinear skbs,
in case the size of contiguous buffer in skb is large
enough to allow recycling. The basic use case is TSO,
when the driver allocates new skbs and then it can add
fragments to the new skb. In this case, skb_is_recycleable()
returns false because skb is not linear.

Signed-off-by: Elena Tebesoi <elena.tebesoi@freescale.com>
Signed-off-by: Claudiu Manoil <claudiu.manoil@freescale.com>
[Kevin: Original patch taken from FSL
QorIQ-SDK-V1.2-SOURCE-20120614-yocto.iso image]
Signed-off-by: Kevin Hao <kexin.hao@windriver.com>
Signed-off-by: Tiejun Chen <tiejun.chen@windriver.com>