powerpc/85xx: fsl-pamu: put the PAACT/OMT in its own coherence subdomain (Erratum A-004510)
Implement the work-around for erratum A-004510 in the Freescale PAMU device
driver.

Erratum A004510 says that under certain load conditions, modified
cache lines can be discarded, causing data corruption.  The work-around
involves multiple steps, one of which is to isolate the PAMU data structures
(PAACT and OMT) from the rest of the system address space before the PAMU is
enabled.

We also need to use a spinlock to ensure that only one core at a time is
reading or writing from these data structures.

The work-around is fully documented in Application Note AN4493.

Signed-off-by: Timur Tabi <timur@freescale.com>
[Kevin: Original patch taken from FSL
QorIQ-SDK-V1.2-SOURCE-20120614-yocto.iso image]
Signed-off-by: Kevin Hao <kexin.hao@windriver.com>