ARM: add an extra pass for kallsyms generation
This error keeps seemingly pop up at random:

 SYSMAP  .tmp_System.map
 Inconsistent kallsyms data
 This is a bug - please report about it
 Try make KALLSYMS_EXTRA_PASS=1 as a workaround
 make[1]: *** [vmlinux] Error 1

The executive summary is that expected differences in symbol
ordering lead to differences in compression.  For full details
see this link:

	http://marc.info/?l=linux-next&m=133267456809502

Implement a temporary workaround so that build coverage is not
sporadically impacted.  This is not to be considered a final
fix worthy of upstream, but just a stop gap measure.

Not-Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>