qman: always build parameter checks for slow-path.
This proved to be a problem when parking held-active FQs during
dequeue-processing. The FQ is not actually known to be parked until the
FQPN notification is processed later on. So APIs were being called that
depended on the FQ being parked, checks were not catching the race
condition that the FQ was sometimes not parked in time, and undefined
results would ensue (particularly when the FQPN *did* show up later on,
and the FQ was longer in a state to expect it).

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>
[Kevin: Original patch taken from FSL
QorIQ-SDK-V1.2-SOURCE-20120614-yocto.iso image]
Signed-off-by: Kevin Hao <kexin.hao@windriver.com>