powerpc/e5500: Separate __restore_cpu_e5500 from __setup_cpu_e5500
Currently restore_cpu_e5500 shares code with setup_cpu_e5500, as
a result we try to operate on the cpu features pointer both on
primary and secondary cores. This is incorrect as the cpu
feature check and handling is only required on the primary core.

Signed-off-by: Varun Sethi <Varun.Sethi@freescale.com>
Signed-off-by: Stuart Yoder <stuart.yoder@freescale.com>
[Kevin: Original patch taken from FSL
QorIQ-SDK-V1.2-SOURCE-20120614-yocto.iso image]
Signed-off-by: Kevin Hao <kexin.hao@windriver.com>