dpaa_eth: Adjust default buffer size
Default buffer size is based on the L2 MAXFRM value, such that
buffers in the pool are large enough to accomodate incoming frames
in a contiguous fd.

But while MAXFRM includes the FCS size, we don't need to account
for that in our buffers, since FCS is stripped down by the DTSEC.
Adjust default buffer size accordingly; it's a minor change, but
which better reflects reality and also helps a bit with the
recycling of termination buffers.

Signed-off-by: Ioana Radulescu <ruxandra.radulescu@freescale.com>
[Kevin: Original patch taken from FSL
QorIQ-SDK-V1.2-SOURCE-20120614-yocto.iso image]
Signed-off-by: Kevin Hao <kexin.hao@windriver.com>