powerpc/booke: Eliminate rfi from exception entry path.
Extracted from QorIQ-DPAA-SDK-v1.1-20111026-systembuilder.iso

Unlike classic, we don't really need the MSR change to be atomic with the
branch.  This eliminates a trap as a KVM guest (in the absence of hardware
hypervisor extensions), where mtmsr is paravirtualized but rfi is not.  For
a virtualized guest without any paravirtualization, this eliminates two
traps (SRR0/1).

Signed-off-by: Scott Wood <scottwood@freescale.com>