powerpc/85xx: fix LAW size calculation in PAMU driver
Fix the calculation of the LAW sizes in the PAMU driver.  It was calculating
a window size that was only half of the actual size, so it only was only
able to find the DDR LAW if the physical address of the PAMU data structures
was located in the lower half of the LAW.

The bug manifests on a 64-bit kernel because that kernel typically allocates
memory at higher physical addresses than 32-bit kernels.

Signed-off-by: Timur Tabi <timur@freescale.com>
[Kevin: Original patch taken from FSL
QorIQ-SDK-V1.2-SOURCE-20120614-yocto.iso image]
Signed-off-by: Kevin Hao <kexin.hao@windriver.com>