DMA/RAIDENGINE: support compute PQ with only one source or one destination.
GenQQ is the only interface we export to async layer  to generate P and Q
parity and this command needs at least two sources and two destinations.

Anyway async layer needs to use this interface to compute Q parity only
or just pass one source to compute Q parity.

If re_jr_prep_pq(...) is called with disable_p, the computation is delegated
to GenQ which computes P or Q only.

If re_jr_prep_pq(...) is called with only one source, a pseuodo source is
given to RE with associated coef all zero which has no impact on the
compuation result.

By this way, this patch mends the interface difference with async layer.

Signed-off-by: Xuelin Shi <b29237@freescale.com>
[Kevin: Original patch taken from FSL
QorIQ-SDK-V1.2-SOURCE-20120614-yocto.iso image]
Signed-off-by: Kevin Hao <kexin.hao@windriver.com>