powerpc/acpx1: add dts for acpx1 board
Copied from lsi_u-boot_4.8.1.62 tarball.

For acpx1 board a dtb is already embedded in the u-boot by default.
But we need to add more device nodes to support other devices, and
we also need to add the partition table for the nand flash. So we
copy the dts file from u-boot tarball, and will adjust the it in
the following patches.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>