introduce-build-runtime-bug-on
Introduce BUILD_RUNTIME_BUG_ON()

BUILD_RUNTIME_BUG_ON - check condition at build (if constant) or runtime

If the condition is a constant and true, the compiler will generate a
build error. If the condition is not constant, a BUG will be triggered
at runtime if the condition is ever true. If the condition is constant
and false, no code is emitted.

Useful to add as sanity check on the input received by API such as
align.h, which typically receive constant parameters which need to be
power of two. Having a build failure is expected when an incorrect
parameter is passed. However, triggering a build failure is not always
appropriate, because the parameter could sometimes be non-constant

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>