DMA/RAIDENGINE: remove annoying error output which is normal case.
Job queue in RE has fix-length of 4096.

Upper level first puts jobs into a list and then move it from list to RE.
If number of jobs is more than RE's size, just stop move job to RE and not need to print "error" message. This is a normal case.

This patch also optimizes the wrap usage of circular array.
We use AND instead of if. Like:
if(count==4096) count=0;
changed to:
count &= ring_size-1;

Changed label to lowercase compared with v1.

Signed-off-by: Xuelin Shi <b29237@freescale.com>
[Kevin: Original patch taken from FSL
QorIQ-SDK-V1.2-SOURCE-20120614-yocto.iso image]
Signed-off-by: Kevin Hao <kexin.hao@windriver.com>