powerpc/e6500: hardware tablewalk support
commit 4844067cdb52da909b08ad717b32e9f448157138 from
https://git.freescale.com/git-private/cgit.cgi/ppc/t4240/linux.git

Preliminary support for e6500 hardware tablewalk.  This is for simulator
(and rev2) only.  On rev1 hardware, be sure to set
CONFIG_PPC_E6500_REV1_BUGS.

Locking is introduced to make TLB misses threadsafe.  TODO: add locking to
the non-tablewalk miss handlers, which was the main reason for making the
locking conditional.

Signed-off-by: Scott Wood <scott@tyr.buserror.net>
Signed-off-by: Andy Fleming <afleming@freescale.net>
[Kevin: Adjust the arguments of macro tlb_prolog_bolted, also fix the
kernel option typo in the commit log.]
Signed-off-by: Kevin Hao <kexin.hao@windriver.com>