powerpc/e6500: hardware tablewalk support
Preliminary support for e6500 hardware tablewalk.  This is for simulator
(and rev2) only.  On rev1 hardware, be sure to set
CONFIG_PPC_FSL_BUGGY_HW_TABLEWALK.

Locking is introduced to make TLB misses threadsafe.  TODO: add locking to
the non-tablewalk miss handlers, which was the main reason for making the
locking conditional.

Signed-off-by: Scott Wood <scott@tyr.buserror.net>
Signed-off-by: Andy Fleming <afleming@freescale.net>
[Kevin: Original patch taken from fsl sdk 1.3.1
QorIQ-SDK-V1.3.1-SOURCE-20121220-yocto.iso.]
Signed-off-by: Kevin Hao <kexin.hao@windriver.com>