I2C: OMAP: Handle error check for pm runtime
commit bc56b0e70ea4e05e59f94f0712eaca130ccebff9 from
git://git.pengutronix.de/git/wsa/linux.git i2c-embedded/for-next

If PM runtime get_sync fails return with the error
so that no further reads/writes goes through the interface.
This will avoid possible abort. Add a error message in case
of failure with the cause of the failure.

Reviewed-by: Kevin Hilman <khilman@ti.com>
Signed-off-by: Shubhrajyoti D <shubhrajyoti@ti.com>
Signed-off-by: Wolfram Sang <w.sang@pengutronix.de>
Integrated-by: Zumeng Chen <zumeng.chen@windriver.com>