netdev/dpa: add the support to adjust the link speed for the dtsec used by userspace
Currently the userspace driver for dtsec doesn't support to adjust
the link speed according to the auto negotiation of phy. So the dtsec
can only works in 1000M mode by default. In order to fix this
limitation, we add the support to adjust the link speed of dtsec by
adding a "fixed-link" property in the mac device node. For example,
if you want to set the dtsec to 100M duplex mode, you can add the
following to the mac device mode.

        fixed-link = <0 1 100 0 0>;

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>
Integrated-by: Tiejun Chen <tiejun.chen@windriver.com>