pch_gbe: fragment packets with length less than 64 fail to transmit
The ethernet driver will check the checksum of UDP/TCP packet with
length less than 64. For fragment packets, the checksum is stored
in the first fragment packet. When ethernet driver check and store
checksum in last fragment packet, the checksum value may be actually
stored in data section of UDP/TCP packet and then destroy the content
of the UDP/TCP packet.

Ethernet driver doesn't need to check the checksum of fragment packets.

Signed-off-by: Jiang Bin <bin.jiang@windriver.com>
Integrated-by: Yang Wei <wei.yang@windriver.com>