pch_gbe: Fix the checksum fill to the error location
Due to some hardware limitations the pch_gbe hardware cannot
calculate checksums when the length of network package is less
than 64 bytes, If a smaller than 64 byte packet is processed, the
destination IP address will be incorrectly modified.

When forwarding network packages at the network layer the IP packages
won't be relayed to the upper transport layer and analyzed there,
consequently, skb->transport_header pointer will be mistakenly remained
the same as that of skb->network_header, resulting in TCP checksum wrongly
filled into the field of destination IP in IP header.

We can fix this issue by manually calculate the offset of the TCP checksum
 and update it accordingly.

Signed-off-by: Zhong Hongbo <hongbo.zhong@windriver.com>
Integrated-by: Yang Wei <wei.yang@windriver.com>