powerpc/e6500: hw tablewalk: fix corruption when no lock needed
If the low bit is not set, that indicates the need for a TLB miss lock,
the next tlb entry was written to the wrong byte, corrupting pgdir_kernel.

This can happen when there's a TLB miss before threading is started
(e.g. serial port MMIO), or if threading is disabled.

Signed-off-by: Scott Wood <scottwood@freescale.com>
[Kevin: Original patch taken from fsl sdk 1.3.1
QorIQ-SDK-V1.3.1-SOURCE-20121220-yocto.iso.]
Signed-off-by: Kevin Hao <kexin.hao@windriver.com>