powerpc/e6500: hw tablewalk: fix corruption when no lock needed
commit 063ad429f86273abd7615487d6accb8ac93c8516 from
https://git.freescale.com/git-private/cgit.cgi/ppc/t4240/linux.git

If the low bit is not set, that indicates the need for a TLB miss lock,
the next tlb entry was written to the wrong byte, corrupting pgdir_kernel.

This can happen when there's a TLB miss before threading is started
(e.g. serial port MMIO), or if threading is disabled.

Signed-off-by: Scott Wood <scottwood@freescale.com>