dpaa_eth: increase number of buffers in bpool
Doubled the number of buffers in bpool and raised the refill level.
This change provides a larger jitter buffer that prevents the
unstable results measured in the forwarding tests while leaving
the termination tests performance unaffected.
Also doubled the taildrop limit to align to the new buffers number.

Signed-off-by: Madalin Bucur <madalin.bucur@freescale.com>
[Kevin: Original patch taken from FSL
QorIQ-SDK-V1.2.1-20120824-yocto.tar.gz tarball.]
Signed-off-by: Kevin Hao <kexin.hao@windriver.com>