LTT & LTTNG2: integration fix
There are 3 fixes for the integration of LTT and LTTNG2:
1. drivers/staging/lttng2/Kconfig fix:
Fix incorrect define for checking if LTT is defined.
(LTT not LTTNG)

2. Cleanup default selections in arch/[x86|powerpc]/Kconfig:
LTT selects by default some defines that are incompatible with LTTNG2:
	HAVE_LTT_DUMP_TABLES
	HAVE_TRACE_CLOCK

3. include/trace/events/timer.h
The top level include/trace/events/timer.h has the same event
definitions as those in lttng2's timer.h.  To complicate matters, the
top level is pulled in first.  Use the top level, comment out the duplcates
in lttng2.

Signed-off-by: Paul Barrette <paul.barrette@windriver.com>