ptrace: scheduling while atomic fix for ARM
The issue at hand here is that the access_process_vm() calls
into get_user_pages which can sometimes call cond_resched().

If ptrace is servicing a breakpoint remove or planting a breakpoint,
a final call out to schedule() should not happen because the system
will be operating in the interrupt context already.

The example here is if ptrace is used to single step every line
of a forked child which exec'ed /bin/ls.  You can end
up with the many lines of errors printed by the kernel:

BUG: scheduling while atomic: ls/0x40000001/641

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>
[Refreshed the context on v3.4 kernel]
Signed-off-by: Yang Shi <yang.shi@windriver.com>