powerpc/law: correct LAW size
The size field, SIZE, of LAWAR dentifies the size of the window from
the starting address. Window size is 2^(SIZE+1) bytes.

So the size really should be calculated as 1ULL << (size + 1).

{
This should be squashed into 87820ec8cbe6a706390ade8fff2172e55c7f95ae

Add a work-around for erratum A-004510 in the Freescale PAMU device driver
}

Signed-off-by: Tiejun Chen <tiejun.chen@windriver.com>