fmd: Modify compat pointer mapping for Fman.
This change affects how the FMD maintains its internal mapping of
64-bit kernel space pointers to the 32-bit so-called "handles"
that are being returned to user space on behalf of those pointers,
in the 64-bit kernel space, 32-bit user space scenario. It does not
apply to 32bit/32bit and 64bit/64bit kernel space/user space scenarios.

The mappings are intended to be stored in pools, one pool per relevant
FMD resource, e.g.: PCD, policer, keygen, etc.; such a userspace handle
is thus irrelevant all by itself, outside of the context of the FMan
resource it pertains to. The context is inferred from the API that the
handle is used within, and userspace needs not be concerned with it.

Although currently only one such pool is implemented, the existing
code employed a char* to identify such mapping pools. That char*
argument to the compat_*_ptr2id() and compat_*_id2ptr() functions
was not interpreted as a string, but rather it was being compared
by value with existing mappings. That not only was misleading, but
was also causing visibility issues for other kernel-space entities
that required access to this API.

This patch fixes the issues created by the use of the char* argument
by replacing it with an enum.

Signed-off-by: Andrei Pistirica <sorin.pistirica@freescale.com>
Signed-off-by: Stefan Szabo <szbs001@freescale.com>
[Kevin: Original patch taken from FSL
QorIQ-SDK-V1.3-SOURCE-20121114-yocto.iso tarball.]
Signed-off-by: Kevin Hao <kexin.hao@windriver.com>