SPEAr: add one memory bank to meminfo in machines' fixup
DB9000 needs to reserve memory for framebuffer, but in fixup, meminfo has
no memory bank info yet. So add memory bank info before reserving the memory.

With this patch, we can't pass more than 1008M number to mem argument in the
bootargs.

In st git tree, there is one commit to work around this issue:
commit 0d083ab7 ARM init workaround: move fixup call after parse_cmdline

This commit modifies common code arch/arm/kernel/setup.c to postpone
calling fixup, but this is not compatible with multi board support, so
following the approach of other boards that use explicit memory bank
info in the fixup is used here.

Signed-off-by: Liming Wang <liming.wang@windriver.com>